+++ src/FlowSynx.Application/Features/Workflows/Command/Delete/DeleteWorkflowHandler.cs
@@ -29,13 +29,19 @@ internal class DeleteWorkflowHandler : IRequestHandler<DeleteWorkflowRequest, Re
     {
         try
         {
+            if (string.IsNullOrEmpty(_currentUserService.UserId))
+                throw new FlowSynxException((int)ErrorCode.SecurityAthenticationIsRequired, Resources.Authentication_Access_Denied);
+
             var workflowId = Guid.Parse(request.Id);
             var workflow = await _workflowService.Get(_currentUserService.UserId, workflowId, cancellationToken);
             if (workflow == null)
-                throw new FlowSynxException((int)ErrorCode.WorkflowNotFound, $"The workflow with id '{request.Id}' not found");
+            {
+                var message = string.Format(Resources.Features_Workflow_Delete_WorkflowCouldNotBeFound, request.Id);
+                throw new FlowSynxException((int)ErrorCode.WorkflowNotFound, message);
+            }
 
             await _workflowService.Delete(workflow, cancellationToken);
-            return await Result<Unit>.SuccessAsync(Resources.DeleteConfigHandlerSuccessfullyDeleted);
+            return await Result<Unit>.SuccessAsync(Resources.Feature_Workflow_Delete_DeletedSuccessfully);
         }
         catch (FlowSynxException ex)
         {
